Не работает запись калибровочных значений во флэш

+++ libraries/common/telemetry.h
@@ -34,7 +34,7 @@ extern "C"
 
   typedef struct
   {
-    uint16_t RPM; // Real value
+    uint8_t RPM; // Real value/1000
     uint8_t FuelCapacity; // 0 - 100%
     uint8_t ThrottlePosition; // 0 - 100%
   } PropultionParametersTypeDef;
